CLOUDSTACK-5216. delete volume failed due to Exception: java.lang.Exception" while destroying Vms

+++ plugins/hypervisors/vmware/src/com/cloud/storage/resource/VmwareStorageProcessor.java
@@ -1517,7 +1517,9 @@ public class VmwareStorageProcessor implements StorageProcessor {
                         vmMo.destroy();
 
                         // this.hostService.handleDatastoreAndVmdkDetach(iScsiName, storageHost, storagePort);
-                        this.hostService.removeManagedTargetsFromCluster(managedIqns);
+                        if (managedIqns != null && !managedIqns.isEmpty()) {
+                            this.hostService.removeManagedTargetsFromCluster(managedIqns);
+                        }
 
                         for (NetworkDetails netDetails : networks) {
                             if (netDetails.getGCTag() != null && netDetails.getGCTag().equalsIgnoreCase("true")) {
